+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## WIP
+## 3.0.0 rc 21 (November 7, 2018)
 
 ### Added